# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: DesiPartyHub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vendors:
    get:
      operationId: listVendors
      tags: [vendors]
      summary: List vendors with optional filters
      parameters:
        - name: category
          in: query
          schema:
            type: ["string", "null"]
        - name: city
          in: query
          schema:
            type: ["string", "null"]
        - name: minPrice
          in: query
          schema:
            type: ["integer", "null"]
        - name: maxPrice
          in: query
          schema:
            type: ["integer", "null"]
        - name: minRating
          in: query
          schema:
            type: ["number", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: page
          in: query
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Vendors list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorListResponse"

    post:
      operationId: createVendor
      tags: [vendors]
      summary: Create a new vendor (admin or vendor user)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
...
  /vendors/{id}:
    get:
      operationId: getVendor
      tags: [vendors]
      summary: Get vendor by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vendor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorDetail"

    patch:
      operationId: updateVendor
      tags: [vendors]
      summary: Update vendor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"

    delete:
      operationId: deleteVendor
      tags: [vendors]
      summary: Delete vendor (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # REVIEWS
  /vendors/{id}/reviews:
    get:
      operationId: listVendorReviews
      tags: [reviews]
      summary: Get reviews for a vendor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"

    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  # BOOKINGS
...
  /favorites/{vendorId}:
    post:
      operationId: addFavorite
      tags: [favorites]
      summary: Add vendor to favorites
      parameters:
        - name: vendorId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

    delete:
      operationId: removeFavorite
      tags: [favorites]
      summary: Remove vendor from favorites
      parameters:
        - name: vendorId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # CHECKLIST
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # AUTH
...
    Vendor:
      type: object
      required: [id, name, categoryId, categoryName, city, state, description, priceMin, priceMax, rating, reviewCount, imageUrl, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        categoryId:
          type: integer
        categoryName:
          type: string
        city:
          type: string
        state:
          type: string
        description:
          type: string
        priceMin:
          type: integer
        priceMax:
          type: integer
        rating:
          type: number
        reviewCount:
          type: integer
        imageUrl:
          type: string
        gallery:
          type: array
          items:
            type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        isActive:
          type: boolean
        isFeatured:
          type: boolean
        createdAt:
          type: string
    VendorDetail:
      type: object
      required: [id, name, categoryId, categoryName, city, state, description, priceMin, priceMax, rating, reviewCount, imageUrl, isActive, createdAt, reviews]
      properties:
        id:
          type: integer
        name:
          type: string
        categoryId:
          type: integer
        categoryName:
          type: string
        city:
          type: string
        state:
          type: string
        description:
          type: string
        longDescription:
          type: ["string", "null"]
        priceMin:
          type: integer
        priceMax:
          type: integer
        rating:
          type: number
        reviewCount:
          type: integer
        imageUrl:
          type: string
        gallery:
          type: array
          items:
            type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        isActive:
          type: boolean
        isFeatured:
          type: boolean
        createdAt:
          type: string
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/Review"
    VendorListResponse:
      type: object
      required: [vendors, total, page, limit]
      properties:
        vendors:
          type: array
          items:
            $ref: "#/components/schemas/Vendor"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    VendorInput:
      type: object
      required: [name, categoryId, city, state, description, priceMin, priceMax, imageUrl]
      properties:
        name:
          type: string
        categoryId:
          type: integer
        city:
          type: string
        state:
          type: string
        description:
          type: string
        longDescription:
          type: string
        priceMin:
          type: integer
        priceMax:
          type: integer
        imageUrl:
          type: string
        gallery:
          type: array
          items:
            type: string
        phone:
          type: string
        email:
          type: string
        website:
          type: string
        isFeatured:
          type: boolean
    VendorUpdate:
      type: object
      properties:
        name:
          type: string
        categoryId:
          type: integer
        city:
          type: string
        state:
          type: string
        description:
          type: string
        longDescription:
          type: string
        priceMin:
          type: integer
        priceMax:
          type: integer
        imageUrl:
          type: string
        gallery:
          type: array
          items:
            type: string
        phone:
          type: string
        email:
          type: string
        website:
          type: string
        isActive:
          type: boolean
        isFeatured:
          type: boolean

    # REVIEWS
    Review:
      type: object
      required: [id, vendorId, userId, userName, rating, comment, createdAt]
      properties:
        id:
          type: integer
        vendorId:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        userAvatarUrl:
          type: ["string", "null"]
        rating:
          type: integer
        comment:
          type: string
        createdAt:
          type: string
    ReviewInput:
      type: object
      required: [rating, comment]
      properties:
        rating:
          type: integer
        comment:
          type: string

    # BOOKINGS
...